[US-3] Create tool store endpoint
I've chose not to use FormRequest to validate the user input, to avoid a
new layer of abstraction. Since we will only have one resource publicly
avilable in the API, and it has only a few fields, we can easily
validate it directly in the controller without increasing the
maintenability cost.